% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Heatmaps_feedact_fcts.R
\name{compute.bites.seq}
\alias{compute.bites.seq}
\title{Compute the number of bites per annotated sequence and trophic guilds}
\usage{
compute.bites.seq(day, seq_df, sp_diet)
}
\arguments{
\item{sp_diet}{}
}
\description{
Compute the number of bites per annotated sequence and trophic guilds
}
