% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Merge_cam_fcts.R,
%   R/3_Interaction_guilds_contribution_fcts.R
\name{compute.contrib.feedact}
\alias{compute.contrib.feedact}
\title{Compute the contribution to the feeding_act for one site}
\usage{
compute.contrib.feedact(df_list, sp_FPA_vect, sp_PPA_vect, studied_site)

compute.contrib.feedact(df_list, sp_FPA_vect, sp_PPA_vect, studied_site)
}
\arguments{
\item{df_list}{a list containing the grazing dataframe of the studied
site with one dataframe for each day}

\item{sp_FPA_vect}{a vector containing the name of the species seen on
the FPA site}

\item{sp_PPA_vect}{a vector containing the name of the species seen on
the PPA site}

\item{studied_site}{the name of the type of site studied: either "FPA" or
"PPA"}

\item{all_sp_FPA_vect}{a vector containing the name of the species seen on
the FPA site}

\item{all_sp_PPA_vect}{a vector containing the name of the species seen on
the PPA site}
}
\value{
two dataframes with the bite rate per species per day (site info) and
the mean bite rate over the days

two dataframes with the bite rate per species per day (site info) and
the mean bite rate over the days
}
\description{
Compute the contribution to the feeding_act for one site

Compute the contribution to the feeding activity process for one site
}
