% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Heatmaps_feedact_fcts.R
\name{data.sequences.scale}
\alias{data.sequences.scale}
\title{Gathers data at the sequence scale when annotations are not continuous}
\usage{
data.sequences.scale(
  start_times,
  end_times,
  seq_length,
  no_seq_length,
  bites_df
)
}
\arguments{
\item{bites_df}{}
}
\description{
Gathers data at the sequence scale when annotations are not continuous
}
