% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Get_species_nm_fcts.R
\name{get.sp.nm}
\alias{get.sp.nm}
\title{Retrieve the species names in the column of the presence-absence dataframes}
\usage{
get.sp.nm(df_list)
}
\arguments{
\item{df_vect}{a vector of the presence-absence dataframes}
}
\value{
a vector containing the names of all the species present in the
presence-absence datfarames columns
}
\description{
Retrieve the species names in the column of the presence-absence dataframes
}
