% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Merge_cam_fcts.R
\name{merge.cam.vid.df}
\alias{merge.cam.vid.df}
\title{Merge dataframes from LD and SD camera for a given day and correct names for
AcCtendark (-> C. striatus) and Scarus sordidus (-> Chlorurus)and remove
all planktivore species except H. fasciatus (wrong annotation)}
\usage{
\method{merge}{cam.vid.df}(list_LD_dfs, list_SD_dfs, start_time, end_time, herb_nm_all_vect)
}
\arguments{
\item{list_LD_dfs}{a list containing the grazingdataframe
from the long-duration cameras ie one dataframe per camera}

\item{list_SD_dfs}{a list containing the grazing dataframe
from the short-duration cameras ie one dataframe per camera and per pose}

\item{start_time}{a time in a character format like "hh:mm:ss" to tell when
the final dataframe must start}

\item{end_time}{a time in a character format like "hh:mm:ss" to tell when
the final dataframe must end}
}
\value{
a presence-absence dataframe with the data gathered from all the
dataframes (long and short duration): time is continuous from the start_hour
to the end_hour even if short-duration dataframes only have annotation on
certain time spans.
}
\description{
Merge dataframes from LD and SD camera for a given day and correct names for
AcCtendark (-> C. striatus) and Scarus sordidus (-> Chlorurus)and remove
all planktivore species except H. fasciatus (wrong annotation)
}
