% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Heatmaps_feedact_fcts.R
\name{plot.heatmaps.act.intensity}
\alias{plot.heatmaps.act.intensity}
\title{Plot heatmaps of activity through time with one heatmap per trophic guild}
\usage{
\method{plot}{heatmaps.act.intensity}(bites_seq_df_list)
}
\arguments{
\item{color_palette}{}
}
\description{
Plot heatmaps of activity through time with one heatmap per trophic guild
}
