% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_Interaction_guilds_contribution_fcts.R
\name{plot.interact.guild.contrib}
\alias{plot.interact.guild.contrib}
\title{Title}
\usage{
\method{plot}{interact.guild.contrib}(contrib_FPA_df, contrib_PPA_df, guild_nm, sites_colors)
}
\arguments{
\item{contrib_FPA_df}{the dataframe showing the contribution of herbivores on
the FPA}

\item{contrib_PPA_df}{the dataframe showing the contribution of herbivores on
the PPA}

\item{scale}{a caracter string either "total" if total herbivory contribution
is to be plotted or "day" if daily herbivory contribution is to be plotted}
}
\value{
a barplot showing species contribution to herbivory for the two
protection level
}
\description{
Title
}
