% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_Temporal_turnover_feedact_fct.R
\name{subset.df.feedact}
\alias{subset.df.feedact}
\title{Title}
\usage{
\method{subset}{df.feedact}(bites_seq_df, sp_feedact_nm_vect)
}
\arguments{
\item{bites_seq_df}{a dataframe with species bites through sequences}

\item{sp_feedact_nm_vect}{a vector containing te name of the feedact to study}
}
\value{
a dataframe with only the species to keep (those of the studied
feedact)
}
\description{
Title
}
